yet to fix optimization

+++ kfusion/include/kernels.h
@@ -145,11 +145,11 @@ public:
 		float zt = pose.data[2].w - _initPose.z;
 		return (make_float3(xt, yt, zt));
 	}
-	void computeFrame(const ushort * inputDepth, const uint2 inputSize,
+	void computeFrame(const float * inputDepth, const uint2 inputSize,
 			float4 k, uint integration_rate, uint tracking_rate,
 			  float icp_threshold, float mu, const uint frame) ;
 
-	bool preprocessing(const ushort * inputDepth, const uint2 inputSize);
+	bool preprocessing(const float * inputDepth, const uint2 inputSize);
 	bool tracking(float4 k, float icp_threshold, uint tracking_rate, uint frame);
 	bool raycasting(float4 k, float mu, uint frame);
 	bool integration(float4 k, uint integration_rate, float mu, uint frame);
